Remove computeHoursForADay from momentUtils
